requirements.txt
pytest==8.3.5
pytest-asyncio==0.25.3
black
pylint
uvicorn==0.34.0
fastapi[standard]==0.115.11
sqlmodel==0.0.23
aiomysql==0.2.0
greenlet==3.1.1
pyyaml==6.0.2
httpx==0.28.1
testcontainers-mysql
dependency_injector
